Added ubos-state functionality and callbacks Removed site templates directory from main disk; not used Refactored staff-based site templates mechanism Refactored site JSON checking and error reporting

+++ ubos-admin/vendor_perl/UBOS/Networking/NetConfigUtils.pm
@@ -790,7 +790,9 @@ END
         my $callbackContent = 'UBOS::HostnameCallbacks::UpdateEtcHosts ' . join( ' ', @appNics ) . "\n";
         UBOS::Utils::saveFile( '/etc/ubos/hostname-callbacks/etchosts', $callbackContent );
     } else {
-        UBOS::Utils::deleteFile( '/etc/ubos/hostname-callbacks/etchosts' );
+        if( -e '/etc/ubos/hostname-callbacks/etchosts' ) {
+            UBOS::Utils::deleteFile( '/etc/ubos/hostname-callbacks/etchosts' );
+        }
     }
 
     # Start / stop / restart / enable / disable services
